+++ src/promptflow/promptflow/azure/operations/_run_operations.py
@@ -50,6 +50,7 @@ from promptflow._sdk.entities import Run
 from promptflow._utils.async_utils import async_run_allowing_running_loop
 from promptflow._utils.flow_utils import get_flow_lineage_id
 from promptflow._utils.logger_utils import get_cli_sdk_logger
+from promptflow._utils.yaml_utils import load_yaml
 from promptflow.azure._constants._flow import AUTOMATIC_RUNTIME, AUTOMATIC_RUNTIME_NAME, CLOUD_RUNS_PAGE_SIZE
 from promptflow.azure._load_functions import load_flow
 from promptflow.azure._restclient.flow_service_caller import FlowServiceCaller
@@ -715,7 +716,10 @@ class RunOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
             # fall back to unknown user when failed to get credential.
             user_alias = "unknown_user"
         flow_id = get_flow_lineage_id(flow_dir=flow)
-        session_id = f"{user_alias}_{flow_id}"
+        # for different environment, use different session id to avoid image cache
+        env = self._resolve_environment(run=flow)
+        env_hash = hashlib.sha256(json.dumps(env, sort_keys=True).encode()).hexdigest()
+        session_id = f"{user_alias}_{flow_id}_{env_hash}"
         # hash and truncate to avoid the session id getting too long
         # backend has a 64 bit limit for session id.
         # use hexdigest to avoid non-ascii characters in session id
@@ -789,7 +793,32 @@ class RunOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
         )
         print(f"Web View: {portal_url}")
 
-    def _resolve_automatic_runtime(self):
+    @classmethod
+    def _resolve_environment(cls, run):
+        from promptflow._sdk._constants import DAG_FILE_NAME
+        from promptflow.azure._constants._flow import PYTHON_REQUIREMENTS_TXT
+
+        flow = run.flow
+        if os.path.isdir(flow):
+            flow = os.path.join(flow, DAG_FILE_NAME)
+        flow_dict = load_yaml(flow)
+        environment = flow_dict.get("environment", {})
+
+        if not isinstance(environment, dict):
+            raise TypeError(f"environment should be a dict, got {type(environment)} for {environment}")
+        if PYTHON_REQUIREMENTS_TXT in environment:
+            req_path = os.path.join(os.path.dirname(flow), environment[PYTHON_REQUIREMENTS_TXT])
+            if not os.path.exists(req_path):
+                raise FileNotFoundError(
+                    f"File {environment[PYTHON_REQUIREMENTS_TXT]} in environment for flow {flow} not found."
+                )
+            with open(req_path, "r") as f:
+                requirements = f.read().splitlines()
+            environment[PYTHON_REQUIREMENTS_TXT] = requirements
+        return environment
+
+    @classmethod
+    def _resolve_automatic_runtime(cls):
         logger.warning(
             f"You're using {AUTOMATIC_RUNTIME}, if it's first time you're using it, "
             "it may take a while to build runtime and you may see 'NotStarted' status for a while. "