+++ src/promptflow/CHANGELOG.md
@@ -9,6 +9,10 @@
   PF_USE_AZURE_CLI_CREDENTIAL=true
   ```
 
+### Bugs Fixed
+
+- [SDK/CLI][azure] Fixed automatic runtime session id cache when image updated.
+
 ## 1.5.0 (2024.02.06)